renamed directory controllers to handlers
