+++ serve-api/handlers/base.go
@@ -1,0 +1,15 @@
+package handlers
+
+import "github.com/ministryofjustice/serve-opg/serve-api/entity"
+
+type BaseHandler struct {
+	orderRepo entity.OrderRepository
+}
+
+func NewBaseHandler(
+	orderRepo entity.OrderRepository,
+) *BaseHandler {
+	return &BaseHandler{
+		orderRepo: orderRepo,
+	}
+}
